Rename main folder from src to evalhub
